--- src/main/java/blackjack/domain/Card.java
@@ -1,41 +1,0 @@
-package blackjack.domain;
-
-import java.util.Objects;
-
-public class Card {
-    private final Figure figure;
-    private final Number number;
-
-    public Card(Figure figure, Number number) {
-        this.figure = figure;
-        this.number = number;
-    }
-
-    public String getCardName() {
-        return number.getName() + figure.getName();
-    }
-
-    public int getNumberValue() {
-        return number.getValue();
-    }
-
-    public Number getNumber() {
-        return number;
-    }
-
-    public boolean isAce() {
-        return number.isAce();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) return false;
-        Card card = (Card) o;
-        return figure == card.figure && number == card.number;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(figure, number);
-    }
-}
