--- src/main/java/blackjack/domain/Cards.java
@@ -1,28 +1,0 @@
-package blackjack.domain;
-
-import java.util.Collections;
-import java.util.List;
-
-public class Cards {
-    private final List<Card> cards;
-
-    public Cards(List<Card> cards) {
-        this.cards = cards;
-    }
-
-    public void shuffle() {
-        Collections.shuffle(cards);
-    }
-
-    public Card giveFirstCard() {
-        return cards.removeFirst();
-    }
-
-    public int getSize() {
-        return cards.size();
-    }
-
-    public List<Card> getCards() {
-        return cards;
-    }
-}
