--- src/main/java/blackjack/domain/Figure.java
@@ -1,18 +1,0 @@
-package blackjack.domain;
-
-public enum Figure {
-    SPADE("스페이드"),
-    HEART("하트"),
-    CLOVER("클로버"),
-    DIAMOND("다이아몬드");
-
-    private final String name;
-
-    Figure(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
