--- src/main/java/blackjack/domain/GameResults.java
@@ -1,34 +1,0 @@
-package blackjack.domain;
-
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-public class GameResults {
-    private final Map<Player, Integer> gameResults;
-
-    private GameResults(Map<Player, Integer> gameResults) {
-        this.gameResults = gameResults;
-    }
-
-    public static GameResults of(Dealer dealer, Players players) {
-        Map<Player, Integer> results = new LinkedHashMap<>();
-        for (Player player : players.getPlayers()) {
-            GameResult gameResult = GameResult.matchResult(player, dealer);
-            results.put(player, player.calculateProfit(gameResult.getRatio()));
-        }
-        return new GameResults(results);
-    }
-
-    public Map<Player, Integer> getPlayersProfit() {
-        return Collections.unmodifiableMap(gameResults);
-    }
-
-    public int getDealerProfit() {
-        int dealerProfit = 0;
-        for (Integer profit : gameResults.values()) {
-            dealerProfit -= profit;
-        }
-        return dealerProfit;
-    }
-}
