+++ src/main/java/blackjack/domain/card/Figure.java
@@ -1,0 +1,18 @@
+package blackjack.domain.card;
+
+public enum Figure {
+    SPADE("스페이드"),
+    HEART("하트"),
+    CLOVER("클로버"),
+    DIAMOND("다이아몬드");
+
+    private final String name;
+
+    Figure(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
