+++ src/main/java/blackjack/domain/deck/Cards.java
@@ -1,0 +1,30 @@
+package blackjack.domain.deck;
+
+import blackjack.domain.card.Card;
+
+import java.util.Collections;
+import java.util.List;
+
+public class Cards {
+    private final List<Card> cards;
+
+    public Cards(List<Card> cards) {
+        this.cards = cards;
+    }
+
+    public void shuffle() {
+        Collections.shuffle(cards);
+    }
+
+    public Card giveFirstCard() {
+        return cards.removeFirst();
+    }
+
+    public int getSize() {
+        return cards.size();
+    }
+
+    public List<Card> getCards() {
+        return cards;
+    }
+}
