--- src/test/java/blackjack/domain/DealerTest.java
@@ -1,43 +1,0 @@
-package blackjack.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class DealerTest {
-    @Test
-    @DisplayName("딜러의 점수가 16점 이하면 반드시 카드 한장을 받아온다.")
-    void canReceiveCard() {
-        // given
-        Dealer dealer = new Dealer();
-        // when & then
-        assertThat(dealer.canReceive()).isTrue();
-    }
-
-    @Test
-    @DisplayName("딜러의 점수가 17점 이상이면 추가로 받아올 수 없다.")
-    void cantReceiveCard() {
-        // given
-        Dealer dealer = new Dealer();
-        dealer.receiveCard(new Card(Figure.CLOVER, Number.ACE));
-        dealer.receiveCard(new Card(Figure.CLOVER, Number.SEVEN));
-
-        // when & then
-        assertThat(dealer.canReceive()).isFalse();
-    }
-
-    @Test
-    @DisplayName("딜러의 카드 하나를 반환한다")
-    void getFirstCardTest() {
-        // given
-        Dealer dealer = new Dealer();
-        dealer.receiveCard(new Card(Figure.CLOVER, Number.ACE));
-
-        // when
-        String card = dealer.getInitialCards();
-
-        // then
-        assertThat(card).isEqualTo("A클로버");
-    }
-}
