+++ src/test/java/blackjack/domain/hand/ScoreTest.java
@@ -1,0 +1,7 @@
+package blackjack.domain.hand;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class ScoreTest {
+
+}
